Rename ...bunsen.mappings package to ...bunsen.codes.
