+++ bunsen-core/src/main/java/com/cerner/bunsen/ValueSetUdfs.java
@@ -1,6 +1,6 @@
 package com.cerner.bunsen;
 
-import com.cerner.bunsen.mappings.broadcast.BroadcastableValueSets;
+import com.cerner.bunsen.codes.broadcast.BroadcastableValueSets;
 import java.util.ArrayDeque;
 import java.util.Deque;
 import org.apache.spark.api.java.JavaSparkContext;
