+++ bunsen-core/src/main/java/com/cerner/bunsen/codes/broadcast/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Support to broadcast code mappings that are useful to
+ * query or transform FHIR data in Spark. This package is intended for
+ * internal use.
+ */
+package com.cerner.bunsen.codes.broadcast;
