+++ bunsen-core/src/main/java/com/cerner/bunsen/codes/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes for working with FHIR concept maps.
+ */
+package com.cerner.bunsen.codes;
