--- bunsen-core/src/main/java/com/cerner/bunsen/mappings/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Classes for working with FHIR concept maps.
- */
-package com.cerner.bunsen.mappings;
