--- bunsen-core/src/main/java/com/cerner/bunsen/mappings/systems/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Support for importing LOINC and SNOMED terminology systems.
- */
-package com.cerner.bunsen.mappings.systems;
