+++ python/bunsen/codes/snomed.py
@@ -1,0 +1,19 @@
+"""
+Support for importing SNOMED relationship files into Bunsen.
+"""
+
+from bunsen.codes import Hierarchies
+
+def with_relationships(sparkSession, hierarchies, snomed_relationship_path, snomed_version):
+    """
+    Returns a hierarchies instance that includes the SNOMED relationships read
+    from the given location.
+    """
+    snomed = sparkSession._jvm.com.cerner.bunsen.codes.systems.Snomed
+
+    jhierarchies = snomed.withRelationships(sparkSession._jsparkSession,
+                                           hierarchies._jhierarchies,
+                                           snomed_relationship_path,
+                                           snomed_version)
+
+    return Hierarchies(sparkSession, jhierarchies)
